@@ -4,16 +4,18 @@ version: 2.1.0
 homepage: http://github.com/flutter-fast-kit/fk_webview
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+#  sdk: ">=2.12.0 <3.0.0"
+#  flutter: ">=2.0.0"
+  sdk: '>=3.4.1 <4.0.0'
+
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_inappwebview: ^5.3.0
+  flutter_inappwebview: ^6.0.0
   rx_event_bus: ^2.0.1
-  url_launcher: ^6.0.3
-  connectivity: ^3.0.3
+  url_launcher: ^6.2.6
+  connectivity: ^3.0.6
   fk_action_sheet: ^2.0.0
 
 dev_dependencies:
